terminator/drop.rs is just one fn... merge it together with the other terminator stuff
